dev: use exceptions in services

+++ src/services/albumService.php
@@ -5,10 +5,11 @@ namespace services;
 require_once ROOT_DIR . 'models/albumModel.php';
 require_once ROOT_DIR . 'repositories/albumRepository.php';
 require_once ROOT_DIR . 'repositories/userRepository.php';
+require_once ROOT_DIR . 'exceptions/BadRequestException.php';
 
+use exceptions\BadRequestException;
 use models\AlbumModel;
 use repositories\AlbumRepository;
-use repositories\UserRepository;
 
 class AlbumService
 {
@@ -30,7 +31,7 @@ class AlbumService
         $this->albumRepo = AlbumRepository::getInstance();
     }
 
-    function getByAlbumId(int $albumId) : AlbumModel {
+    function getByAlbumId(int $albumId) : ?AlbumModel {
         return $this->albumRepo->getByAlbumId($albumId);
     }
 
@@ -51,11 +52,17 @@ class AlbumService
     }
 
     function updateAlbum(int $album_id, string $title, int $user_id, bool $deleteCover, ?array $coverFile, array $music_ids) :?AlbumModel {
+        if (!$this->getByAlbumId($album_id))
+            throw new BadRequestException("Album id does not exist");
+
         return $this->albumRepo->updateAlbum($album_id, $title, $user_id, $deleteCover, $coverFile, $music_ids);
     }
 
-    function deleteAlbum(int $albumId): bool {
-        return $this->albumRepo->deleteAlbum($albumId);
+    function deleteAlbum(int $albumId): void {
+        if (!$this->getByAlbumId($albumId))
+            throw new BadRequestException("Album id does not exist");
+
+        $this->albumRepo->deleteAlbum($albumId);
     }
     function getByUserID(int $userId, int $page) : array {
         return $this->albumRepo->getByUserId($userId, $page);
